f_(refs DPLAN-16179): adjusted code changes in LocationSearchController.php and DemosPlanProcedureListController.php

+++ demosplan/DemosPlanCoreBundle/Controller/Platform/LocationSearchController.php
@@ -41,7 +41,7 @@ class LocationSearchController extends BaseController
             $limit = $query['maxResults'] ?? 50;
             $result = $locationService->searchLocation($query['query'], $limit);
             $suggestions = [];
-            $maxSuggestions = $query['maxResults'] ?? (is_countable($result) ? count($result) : 0);
+            $maxSuggestions = (int) ($query['maxResults'] ?? (is_countable($result) ? count($result) : 0));
 
             for ($i = 0; $i < $maxSuggestions; ++$i) {
                 if (isset($result[$i])) {
