+++ demosplan/DemosPlanCoreBundle/Controller/Procedure/DemosPlanProcedureListController.php
@@ -494,7 +494,7 @@ class DemosPlanProcedureListController extends DemosPlanProcedureController
             $result = $locationService->searchLocation($requestGet['query'], $limit, $maxExtent);
             $this->profilerStop('searchCity');
 
-            $maxSuggestions = $requestGet['maxResults'] ?? (is_countable($result) ? count($result) : 0);
+            $maxSuggestions = (int) ($requestGet['maxResults'] ?? (is_countable($result) ? count($result) : 0));
             // Es gibt Ergebnisse, aber weniger als maxResults
             if ((is_countable($result) ? count($result) : 0) < $maxSuggestions) {
                 $maxSuggestions = is_countable($result) ? count($result) : 0;
